blas: Use ranges
Simplify BVH nodes by using ranges (offset, count) instead of owning
the vector of primitives. These primitives are stored in the BLAS object.

Moreover, if the node has primitives it means it is a leaf. Otherwise,
the left child index is store in triangles.offset and the right child
index is just the left node index plus one.

+++ rayca-soft/src/bvh/tlas.rs
@@ -196,14 +196,14 @@ impl BlasNode {
 
 pub struct TlasBuilder {
     scene: BvhScene,
-    max_depth: usize,
+    max_depth: u8,
 }
 
 impl Default for TlasBuilder {
     fn default() -> Self {
         Self {
             scene: BvhScene::default(),
-            max_depth: usize::MAX,
+            max_depth: u8::MAX,
         }
     }
 }
@@ -214,7 +214,7 @@ impl TlasBuilder {
         self
     }
 
-    pub fn max_depth(mut self, max_depth: usize) -> Self {
+    pub fn max_depth(mut self, max_depth: u8) -> Self {
         self.max_depth = max_depth;
         self
     }
@@ -245,7 +245,7 @@ impl Tlas {
         TlasBuilder::default()
     }
 
-    pub fn new(scene_draw_info: &SceneDrawInfo, scene: BvhScene, max_depth: usize) -> Self {
+    pub fn new(scene_draw_info: &SceneDrawInfo, scene: BvhScene, max_depth: u8) -> Self {
         let mut ret = Self::default();
 
         let model_count = scene.models.len();
